Hole in ground mechanic almost done
Added "damage_amount" mechanic to th Obstacle and BodyPart class scripts that allow setting obstacles to deal a certain amount of damage (1 or 100, for now). 100 (or "full") is dealt when falling through a hole in the ground. I still need to make a good collision shape that represents the hole in the ground.