Ravine jump attempt, Obstacle slowdown "none" setup
Began hacking together a questionable ravine wall mechanic that stops level movement unless player jumps out. It's really fun so deserves work. Might be better with Y-level based "collision" than actual collision.

Also added proper conditions on bodypart script if an Obstacle's slowdown is none. (This was originally for the ravine top-level being a way to function as an object but that's been left behind now)