Damage flash indicator v1.05
Using recursive functions via timers, lower health (on obstacle collision) causes the affected limb to flash a dark red color. This may not be a best way to approach it. Will be trying a v2. Currently the flashing does not speed up as damage increases, nor is the flashing synced up.