Damage flash indicator v.2.05
Chatter dragon blade helped me setup the flash as a coroutine, bringing the number of functions for the flash effect down from 2 to 1. The flashes sync up, and occur on collision but technically the flash timer is always running.

To solve/workaround a problem of the flash timers not beginning at the right times, I made the speeds of the damage begin and the damage end timers different.

There is an issue of how the flashing will be sped up as current health goes down, but I'll work on that tomorrow.

For now the code is looking much more elegant than the last version, but there is the problem of how the flash time will be changed if part of this is run in the ready function as it is.