Updated Blender Models to undo 4.2 swap attempt
Thanks p3p33 for the suggestion. Godot upgrading the meshes in 4.2 messed them up in a way that GitHub couldn't touch. All we had to do was save over them with the original blend files.

Big props to pierogigon as well, for helping me create a new branch in GitHub that omits the 4.1 vs 4.2 reversion chaos/confusion!!