Hit points and HP Bar reworked as Animation-based
Credit to My2Cents for this idea and implementation, and Mini for inspiring the blinking effect!

Separate functions are called when damage or healing are detected, each firing off respective animations that blink the hit points away or back in.

These animations are duplicates of an original, allowing for multiple hit points to be added or subtracted in sequence and simultaneously, regardless of overlapping "timers"/animation lengths.

The HP bar appears and disappears similar as in the previous version.

The biggest improvement of this version is that there is no discrepancy between how the damage changes are shown when timing is non-uniform.