Added "Heart" body mesh; Green Parts share texture
All of the green body parts now share the same texture; meaning changing one changes them all.

Adjusting anything away from default will disrupt the rudimentary "damage material" to "default material" damage flash mechanic from BodyParts script.

Added an alternate Body mesh that includes a heart embedded in the center. Designed to match the empathy bar. This heart is unshaded.

Adjusted the rotational positions of the walk and run and walkfeed animations to optimize outline around heart. The rotation values are overall more consistent across the animations now.

Discovered a "neutral_bone" that was probably automatically created when the Head and Body were separated. It seems to move the Body in strange ways. I can't find any reference to it in the Blender project, meaning I don't know how to re-weight paint it or whatever. So for right now, the WalkFeed animation rotates around the body instead of the body itself rotating around.

Adjusted the rotation of the "Player" in the main_scene to be at 0,0,0; it was a little off.

Added (now hidden) body shields on each body mesh, that in theory could represent a swirly smoky shader that implies the indestructibleness of the Player's body. It's semi-transparent, but looks terrible and so I have to hide it. But it's useful for planning I guess.