Replacing Player Model & Added Shape Keys
Began the process of replacing the Player Character model, in order to properly use Shape Keys that shrink the character's limbs as they take damage.

Model is technically in T-pose, but existing implementation is not, and so it isn't taking the shape keys' affects correctly, resulting in distorted limbs (arms).

New model is added into the main scene, and the old one is hidden.

Adjusted the collision of the legs so the feet actually touch the ground.

Added a debug option (3) for testing the shape key animation.