Copter Added; Debug Obstacle Attempt; Limbs Hide
Added a new NPC model: Copter, and made it a damaging obstacle. Animated its blades with a scene-nested Animation Player (might want to move that).

It is present on a couple terrain blocks.

I attempted to add toggleable debug obstacles (ie multi-hit trees) but didn't finish setting it up properly.

For now, an exported boolean can be set on the relevant nodes that just move it offscreen when it should be hidden.

Also made the body parts hide after a set amount of time, which is now defined in a variable instead of the previous 1.25 magic number.

This variable also feeds into time before game restarts after death, with .3 added onto it. So only a little magical.

I also adjust the offset dictionary entries for the limbs when they shrink; it's about .0001 further now I think. It's better