Helicopters spawn and approach player
Helicopter script is more separate from the spawner script now.

The helicopters stop before getting to close to the Player.

My2Cents helped me set up the helicopter movement to be based on velocity instead of just position-moving. This is to prepare for using a NavigationAgent3D apparently... That will in theory help the units avoid each other.

Had to flip around the helicopter model in godot so it can use "look_at" to always follow the player via rotation, which apparently inverts the direction.

Added an "empty" dirt block and renamed the "empty" grass one to add more clear movement to the terrain when nothing is on it.