Helicopters Move More; Obstacles Improved; More
Helicopters adjust their positions based on how close they are to the player, and how many have stopped in front of the player so far, leading them to form into the distance. Only 6 spawn at a time. Logic is set up to include different collision areas for different unit types. The collision shape doesn't yet decrease when units are removed because units cannot yet be removed.

Helicopters also have Navigation Agent Nodes that allow them to avoid trees and each other when entering and moving around the scene.

Trees have been outfitted with Navigation Obstacle nodes to make them easier for the copters to avoid. It's okay.

The player got a Nav Obst node too but it's currently disenabled because it wasn't helpful.

Added a block test scene for troubleshooting spawner performance issues.

Spawner performance issues improved by changing spawn instantiation to "preload" instead of "load", and by changing what the helicopters (which are spawned) can collide with.

Overhaul of the collision/mask usages throughout the project saw many, though incomplete, changes to both code and inspector aspects alike.

Removed the debug trees and debug node from the relevant terrain block because it seems like a bad idea.

Attempted to fix the stuck-in-cliff issue by changing the ground type to rigidbody but it didn't help. The rigidbody changes remain however.

Attempted to adjust what the camera raycast can see but didn't follow through with that idea; may not actually have done much. Collision_Mask =3 added to the ray_query for now though. In theory it could help us destroy obstacles.

Collision overhaul mentioned earlier includes renaming layers and masks, such that mask 3 is for obstacles only.

Layer/Mask 16 is now, arguably inexplicably, representative of the Player. Sorry

Replaced most of the duplicated Area3D obstacles on only a couple terrain blocks with fully instantiated scenes so I could add the navigation obstacle nodes in. This needs to be done for every obstacle but isn't yet. Improved the naming conventions of these obstacles.

I attempted to add a height limiter in for the helicopters but I don't know what I'm doing so commented it out.

I renamed all the terrain blocks to be better. Created a true "empty" block.

Is that it? Probably not. My god this project's tasks and complexities are ballooning.