Adjusted grab detect scripts; fixed copter collide
Copter collision got broken in the last update; fixed it by improving the naming and collision layer of the copter top level node.

Adjusted the grab detection signal-sending situation to be received by relevant objects (copter and health), and then determined if that it is the object that was hovered/grabbed.

Helicopters can be "destroyed" if clicked on.