Arms and Head follow Cursor + Raycast improvements
Raycast script on Camera3D is much more optimized and eliminates leftover bits that used to make sense.

Misc other changes to layers and things.

Biggest change is that the arms and head now follow the cursor position.

Created a cursor zone in the main scene to detect cursor position.

Made adjustments to the animations of the Player. Currently the arms do not animate anymore for walk or run. Also, the RESET animation has been changed to T-Pose, essentially.