Arm grab attempts exist; nothing perfect found yet
My2Cents helped me set up possible ways to get the alien's arms reaching out to smack helicopters.

A couple different approaches exist on the Alien/Player script, one of which is commented out.

The commented out approach may ultimately be best, but it needs work, and involves making the arm rotate to the direction of the helicopter, and ultimately stretching to approximately the helicopter's distance. This method is still stuck in the rotation stage, but it's getting there.

As for the currently implemented method, it involves moving a bone's position to approximately the helicopter's position. It both sort of and sort of doesn't work. It indeed jumps to the copter, and retracts, but it stretches the arm undesirably and basically uncontrollably to do so, resulting in an unsatisfactory appearance.

I began work on an animation and condition for it that served as an alternate walk cycle without arm movement, but it became unnecessary due to the use of bone position overrides. I'm keeping it around for now though just in case.