Attempt to tween arm pos towards helis
Set up a modified "aim to target" function that uses a transform matrix to point an arm bone at its intended target.

Works well, but when combined with a tween_method on a child bone, it doesn't align properly to the target anymore.

Added some debug stuff to ensure the arm is moving correctly under certain conditions.

Alternative method will be to scale the same arm bone that rotates in the override area.