Arm tweens to heli position; added a fence
My2Cents (and DB a bit!) helped me set up the arm rotation to properly point at desired objects, and then stretch out to reach them (thanks to Mini for helping with that too).

It tweens out but doesn't tween back in after I set up the helicopter getting "freed" post hit.

I added a fence model and experimented a bit with pieces Left, Right, and Post.

Added a messenger signal for "ungrabbing" which signals to the copter that it's "been hit". This will need some adjusting.