Copter grab improvements; Animation setbacks
Helicopters can be uniquely attacked without seemingly impacting other helicopters who may also be detected.

Arm retracts correctly after an attack.

Added a bunch of debug prints on the raycast, player, and copter to help solve that.

Added an AnimationTree and some new code to attempt wedging a grab/attack animation in, but the AnimationTree blending is glitching.

In trying to solve this, removed the Reset animation, and this needs to be replaced.