Animation Stuff Is Good!
Blender to Godot pipeline is better understood and for now, experiments are over.

The best way to avoid conflicts between animations already existing in Godot IF those animations don't have the same tracks animated is to insert keys for all bones on the animations that are missing tracks present in other places.

I added an updated RunsWalk cycle image to the project files in case anyone wants to use it, even if it's not perfect haha, it might help someone (the toughest part is remembering the repeat iconned ones mean that the legs/arms need to flipped position-wise, because the original drawing didn't include the full cycle...)

Not much else changed here otherwise! Most work was done in Blender.

Oh actually I applied transforms of the player model which moved the model into a less than ideal position (ie a little too low relative to the existing walk/run cycle animations, so it might need to be adjust in terms of how it sits on the terrain); but I'll worry about that later when the new walk/run cycles are complete.