Objects can be destroyed
Updated the Hitpoints script to calculate damage taken percentage and apply that to an animation player.

The animation player of each object will need to uniquely hide layers of different objects to reveal their innards.

The house is setup the best for this so far.

The helicopter is also technically setup but not with much care. It was just to make sure it was scalable. It is!

I realized that my previous change to objects, which made their root nodes Area3Ds, was a mistake. That's because it clears the inheritance with its original mesh file, thus making the objects un-updateable from Blender.

So the new solution is present on a newly made House scene. This scene has its original Node3D as the root, and the Area3D for the obstacle is a direct child, with a collision contained inside. Then the Hitpoints Node3D scene is a child of the Obstacle.

This preserves the inheritance. It will need to be applied to other objects as well. Hopefully it doesn't cause unforeseen breakages. Luckily the Hitpoints script just looks one level above it for the Obstacle node's variables/signal, thus there is no issue I can see for now.

I made a new terrain folder for debug, that is being used to test the destructible house. The Terrain Controller is set to load this terrain_debug folder instead of terrain_blocks.