Added exploding houses
Houses can be exploded via particle effects that emit along with the Animation_Degrade animation player set up previously.

Made some changes to the Hitpoints and Obstacle scripts so they communicate a little better with each other, including a health_effects function that ensures certain things happen to the respective Obstacle/Copter and Hitpoints nodes when the health reaches 0, eg, removing collision and damage labels.

This was done so that the particle effect can continue after the object is "destroyed"; previously, the entire nodetree was removed via queue_free but now that is delayed by the length of the particle effect lifetime; the "is dead" aspect of "health_effects" fires sooner so that the object can essentially be dead without actually being dead (until the particle effect lifetime ends).

The actual lifetime of every particle is probably longer but it shouldn't matter for now.