Added a layer of destructibility to the helicopter
Tried to make the helicopter scene NOT have an area3d as the top level node, and I couldn't quickly get it working properly.

The reason why a Node3D is preferable as the top level is that it makes updating the mesh in Blender possible, because it retains the original inheritance.

But that's only true if the scene is created FROM the .blend file.

If I take the original Area3D scene and just drop an inherited version of the mesh in there, then it updates correctly, and I don't have to mess with Node3D being the new top level.

But, this might not be as editable as if it was morphed into the scene directly from the .blend file, idk. But it worked well enough that I could animate the meshes with an animation player (as long as I set the inherited scene to have editable children), so I think it's all fine.

Anyway, that's why there's a new copter scene. It's not being used but it exists now. And then the old helicopter scene now has some changes.

All in all, pretty much, the only change was the copter's windows can be broken by swapping a broken windows mesh in. No particle effects yet.

Also made some debug print message changes that I probably should change back actually. Oh well. This commit is long enough.