Helicopter destruction seems good enough and done
Helicopter spins out and crashes.

Added sub-obstacle that activates after the copter "dies" with its own hitpoint so that it can be attacked while it's careening towards to player, protecting the player from damage that will otherwise be dealt on contact.

Returned the terrain blocks back to previous arrangement.

HIT_DELAY is gone for good.

Added a few "error" and breakpoint messages/conditions on the hitpoint and obstacle scripts that make sure they're setup correctly.