Implemented more workable cow grabbing
DragonBlade shared some code that made picking up the cows work so much better.

It has revealed many other things that also need to be put in place for it to function correctly, but now clicking and dragging the cow around does not result in dropping!

I made a change that keeps it on the Z axis it was originally clicked at. I think.

I added some spawning cows that I can't seem to get to spawn at the position I want but they sure do build up elsewhere.

Clicking on multiple cows results in the cow teleportation problem.

Made other adjustments to the cow rigidbody that sets the velocity to 0 when it's released.

Gravity scale was increased on the cow RigidBody as well that allows it to fall better.