Added working score dunk!
Orb descends from sky when meat is grabbed, and retracts when meat is dropped.

On hover with grabbed object, the orb will jiggle.

If the meat is dropped while hovering over the orb, it will be abducted, and carried away. A point will be registered.

Needs some adjustments but the system works!

I also moved the character mesh and collisions closer to the center of the scene.

Made some slight adjustments to the house particle effects.