Score Dunk improvements
Score dunk moves in and then out based on a timer that increases if you grab more meats.

Started to add a combo system based on how many meats you add in before it retracts but decided against it for the time being.

Abductions are constantly tallied in the process function to in theory be added as a "+1 +2 +3!" etc popup, left unfinished as label3d under the score dunk. This would then be calculated into the abduction score after the retracting completes.

Adjust score dunk appearance and position. Meat inside is a little lower in there now.

Set up a system that makes the dunk more transparent when the mouse is closer to it but quickly deprecated it coz it looks bad. It's just commented out.

Fixed some collision issues between the meats and the Player. Meats now grab to the z axis of the player, and the score dunk is also aligned with this more or less.