Prepared and setup the Cowfarm terrain scene
Added hitpoint stuff, obstacle stuff, arrows, etc to the objects in the cow farm terrain scene.

Decided that fence posts will not be obstacles and just stay there after the fences are destroyed because why not.

Old cows have all been removed. New ones not added yet.

Animation_Degrades are added to the hitpointstuffs on each obstacle BUT the "degrade" animations are not yet created, so the destruction function doesn't play.