Meat objects will reparent as necessary
When grabbed, meat objects reparent to the main scene.

When ungrabbed, and after hitting a surface on a terrain block, the re-reparent to the node that surface itself is parented to.

Made various naming convention changes that arguably do not matter.

Finally made good use of the meat objects' raycast detect surfaces.