New terrain controller is implemented!
ThotTee helped me set it up and it works pretty well!

A few issues are present, including a terrain block that follows the player, and physics objects that build up and stack on each other.

But the lag spikes are now gone and it is definitely a more optimal way to load in scenes!