Various tasks
Started out making a texture for ground meshes.

Redesigned obstacles and stuff on level one.

Fences no longer do damage to the player.

Added a delay boolean for objects that take damage. Ie, it can be set to either need a delay or not, ie if the player goes to attack it, then it would need a delay, whereas if collided into, it gets no delay. This is set as an argument when the damage detection signal is emitted.

Now, non-damaging obstacles (like cornfields and fences) will be destroyed immediately on collision, instead of waiting for the player attack to finish (ie no more delay).

Fences now have color textures that can be loaded in and added to particle effects. Particle effects have been added to fence_01. Fence_01 has a proper-er row 5 version.

Adjusted camera position so the cornfield and fence destruction could be more visible.

Score Dunk position needed to be adjusted to fit new camera spot. Refactored some of the score dunk code, and cleaned it up a tad.

Meat objects no longer disappear off camera since it moved, so I set up the meat objects to reparent to the block below them using the raycasts. Fixed and cleaned up that code, and used a group "Dropped" instead of setting a flag. No problems now.

Adjusted the damage flash timer code to NOT flash when a non-damaged obstacle hits the player after the player has taken some damage. This also led to refactoring the damage flash timer code.

Probably other stuff maybe.