Texture, Material, and mesh-only stuff.
Set up a texture import method for adding textures to meshes. It is not finished and needs to be set to proper colors besides black and white. Visible on cattle ranch 01 on level 01 block.

Saved materials for tree 01, house 01, and the fences that can be loaded in elsewhere for eg particle effects (and are now matching for all the particle effects for the aforementioned objects).

Mesh-only variants created for house 01 and tree 01.

Possibly resolved (or workedaround) an issue where potentially-collidable meat objects would return an unexpected but fatal error seemingly randomly by making sure that a collider var is not set in Camera3D without the raycast result having a collider on it first.