Added a new ranch block
Experimented with lighting stuff for optimization purposes. Didn't turn out to be used.

Added cattle ranch 02 with its own textures on it.

Optimization experiments with cornfields were attempted and abandoned except for LOD being lowered considerably. It helped a lot with performance.

Made a "mesh only" fence scene; it is for the full row.