Added a tool shed asset
Set and saved a couple of materials from the cow barn and reused them for the tool shed.

Updated, exported (1) and reused some particle effects from the house for the tool shed.

Turned off "Disable Z" from a couple particle fx on the house, and all on the tool shed, which resolves an issue with how the particles spread.

Created a tool shed scene, set it up as an obstacle proper, and added a couple to each currently made ranch scene.