Particle FX Fix Attempt
Thanks to JoeSchmoe the particle fx on the mounted solar panels no longer continue playing after they are "unloaded".

This entailed adding a method call animation track to Animation Degrade "degrade", which restart()s the particle effect at the beginning of the animation.

I began adding a check for this track to exist on every degrade animation, but didn't finish setting that up.

Re-refactored the something_hit in the case of actual grab attacks, it is now known as something_attacked.

Adjusted the cow ranch shader.

Locked down the terrain block instance block_blank-grass

That's probably it