Terrain chunks get collected later
Particle FX restart themselves upon getting added to the terrain collector and also last longer because the terrain chunks no longer move when their position is /2 of their size.

Corn stalks now have a particle effect.

Removed the method call from animation degrades that was restarting the particle fx in the animation itself. Stopped working on a manual error that would come up if the method wasn't present.

Toyed with the idea of instanced particle effects but abandoned it after realizing how complex the workflow would ultimately be.

Moved the terrain collector back further. Increased the number of terrain chunks that spawn in from 8 to 9.