Added a windmill asset/scene
Created an animation for the wheel and tail spins. Added the asset to ranch 02. It is not yet setup as a proper obstacle, and still uses its default Blender materials.

I also shrunk the electric pole scene instances down by 25% in each ranch scene.

NPC Spawner is currently no longer setup as an autoload, so  helicopters do not spawn.