Chunks pull from 3 Weighted Terrain Collectors
Precipire helped me setup the 3 collector types so the chunks won't try to be reparented to the controller if they don't exist, and also got them to be added back to their correct collectors.

Finally, they also helped me setup the RNG/weighting so that there is a 50% chance that safe chunks spawn in, and 25% chance that either the obstacle or point chunks spawn in.

It works fabulously.

I moved a floating tree down in the forests also.