Chunks playlist partially implemented
Code from DragonBlade has allowed for setup chunk playlists to be fed through an array and presumably add the list to the terrain controller.

The order of the chunks in the game are incorrect however, so needs some work.

Albesca helped sort out the calls to the function in progress_terrain