Performance improvement, better setup
Nuked all of the Navigation nodes that I could find in the current level. Might have missed some but probably not, but maybe.

That change has increased performance tremendously, although the ranches still lead to a bit of a frame drop when they appear on the controller.

Setup the level folders to have enough chunks, although they are mostly duplicates. All the debug chunks are put away.

Removed the second forest obstacle scene and replaced the first one's trees with the actual latest tree.