Safe level chunks start the level!
Set up system for detecting whether enough safe chunks have gone by at the start, and if so, then load the level chunks.

Right now it does look at the current level to decide which ones to bring in, and it's technically scalable by manually setting each level (1,2,3) to be associated with the blank empty lists in chunk_chosen_to_add, but someday I'd like to make it append the level number (_level_NUMBER) to update it instead, so it's more easily scalable.