Added a teleporting arrow
A few obstacles and cows have it. There are two arrows in the hover arrow scene (a new one) so that one can be on top and the other is behind assuming it doesn't look right due to the camera fov blur bug situation in godot.

Made a good version of tree_03 and replaced it where necessary