Obstacle hilite is considerably more efficient
Thanks to tekhium's suggestion and guidance, and help from ThotTee, the obstacle hilite is more efficient.

It was improved by creating an array that collects the obstacle meshes that are hilited, which are then removed from that array when no longer hovered.

Additionally, a "hovered_old" variable was created to ensure that neither too many nor too few obstacles were being hilited at a time.