Meeting abduction minimum = powerups > next level
There is an abduction minimum that when met, stops the game to drop down some powerup orbs which can be clicked on. When clicked on, the game will continue but on the next level.

Only works when going from level 0 to level 1, for now.

Made changes to the debug menu that allows for adding points.