Scaled level progression all the way to 12
Using cheats, the Player can now get all the way to the "end of the game" (level 12).

Added so many placeholder levels in, to account for every possible level in the game.

Beating level 12 leads to an error due to level 13 not existing.

Added the world enviro back in. Commented out some debug things.

Created some gradients. Decided on terrain_textures as the main gradients folder but need to consolidate this better and get rid of the additional textures folders (like for level 1 and 7)