Began setting up menu to game transition
Right now debug menu is only way to swap into the game mode, but it works and starts the game.

There's an enum on the terrain controller that is updated by the debug buttons, and this changes what function runs to make the game work.

This is currently done by a duplicated function for progress_terrain. It's messy and not ideal. There is a minimal difference between progress terrain_menu and prgoress_terrain_game; one goes infinitely, and the other will end and start the game.