Main menu level WIP
Setup an enum state machine on the terrain controller that changes to a different progress_terrain type depending on what type of level we're supposed to be on.

Currently if set to MENU, it will load in an infinite black-chunk'ed level.

Using debug tools, we can switch to the "game" type, and it will load in the game.

But this only works if the current global level is set to 0. As soon as it's set to 1 or more, chunk_add becomes null.

Next steps are to get the main menu to be able to blend into level 1 instead of the debug level.