States adjustment for terrain controller
now there's two states one for debug and one for game. Menu is mixed into game.

Made other changes that are relevant to having the menu be workable with the game and stuff.

Unfortunately, any chunks that are not menu chunks take a long time to blend away via queue free on level update