"Postmenu" and "Prebegin" gamestates defined
Two new game states PostMenu and PreBegin now work and have special actions.

Essentially, choosing "a character" will abduct an alien in the main menu, and teleport than alien into the sky. The camera follows. This is the postmenu state.

Then the prebegin state has the camera swooping down and starting the game. In here I will want to add another animation.

Made other changes like the score dunk disappears when it is offscreen, and reappears when it is on.

Saved a material for the earth-lit orbs vs the alien-planet-lit orbs.

HUD loading screen before the game starts is currently turned off.