Setup new states, SOLVED terrain overspawning issue
Terrain overspawning issue was due to chunks_list_current not being empty at the time of level_update.emit(1). Setting it to clear on the prebegin state resolves the issue.

Play state now happens after the alien teleport animation plays. Full animation is not complete but it does play. Other stuff like that. Moved the loading text to be part of the blackout for purposes of canvas layering order; loading now appears during the prebegin blackout.