Body limb improvements
Limbs have 2 hp each, including head.

They swap to a smaller mesh when on the 1 hp state, otherwise they hide.

There are now smaller collisions that are swapped to so the player falls down to half-high position. They also fall all the way down when at 0.

Adding healing debug buttons that reinsert the legs and the collisions. Attempted to setup tweens that will increase the player height as well. It works good enough but has a bit of jank.

Something that will help is making sure the second legs don't run the tween when healing.

Head rotation is back. That might have been an earlier commit, I don't remeber.