Gamepad "support", functional healing system
By hovering over the player while an Abductee is grabbed, the Abductee is "consumed", and heals are applied +1 to every limb.

Adjusted the scoredunk ascension to happen sooner if Abductee is consumed.

Began crude gamepad support and swapped all movement controls to no longer be dependent on ui inputs.

Gamepad support works well for movement and grabbing but the crosshair for aiming (right stick) is very choppy. Not sure why.

Setup an alternative grab and drag system for Abductees that is currently not enabled but attempts to shift the position of the abductee up, down, left, or right (up currently doesn't work) to an offset position based on where the cursor has moved after the initial grab. It doesn't feel great but could work with some tweaking.

Setting up Gamepad support was supposed to help decide what method of post-grab movement felt best.