Alien feed animation and collision set up
There is a special area on the head of player (v3) that detects if an edible thing is held over it.

The player_hover signal now reports "true" when seeing the CharacterBody collisions (which don't include a head collision) OR the Head Area collision.

Feeding animation is added in as a oneshot and implemented to occur after the feeding begins. Physics process priority for the player's animation tree is adjusted during feeding time so that looking doesn't override it.

Removed the "Cheat Menu" column from the debug because the healing buttons don't work/are outdated.

Unrelated, made mask 3 collision on the limb areas be set to false, and layer 16 be set to true in the ready function.

Mask 3 was set to true in the inspector (as an "Obstacle") but this seems unnecessary, so I made it false via code. It's still true in the inspector. Doesn't seem to be breaking anything.