Limb collisions improved
Limb collision improvements, including an attempt to ensure that limbs can only be damaged by the same obstacle once at a time.

Added a new body part for the torso so that things can be destroyed even if they don't hurt the player. This necessitated a lot of if statement caveating that I don't feel like describing but it's in this commit haha.