Arm attacks no longer hurt
Refactored signals and functions on the player script.

Renamed attack functions from "grab" namings to be attack namings.

Attacking now turns off the collision of the right arm, and only occurs if "attack" is false. Attack becomes true at the beginning of an attack.

After the attack is finished, the collision returns, and attack becomes false again.

This does not yet account for multiple arms.