Pause Menu and State work!
pressing enter/escape/button6 pauses the game if the "play" state is true, and if in "pause", then the game unpauses when escape or button6 pressed. The continue button is prefocused on release of pause button, leading to enter to already unpause it.

Changed the debug canvas layer to be below the pause menu.

Designed the pause menu, added code. Only functioning button atm is continue.