Restart/Main Menu buttons work!
Game can be restarted via any of the main menu buttons, which are renamed to Restart when you get a game over.

Retry has been deprecated due to how much work it will take to implement. It may be reintroduced at a later time, as a stretch goal or something.