Farmhouse added; big grain bin; shadow changes
Farmhouse now added into a scene and placed on cattle ranch 02. Grain bin is bigger now. Adjusted the shadow bias on the directional light, it caused a very subtle bar of shadow at the bottom of the screen by the player's feet but other shadows on objects look better.

Rename human files a bit.