Abductees drop when held outside interact zone
Code is setup on the interact body entered/exited script, somewhat duplicating the dropped code that exists on the abductees themselves.

Also re-introduced the pit falling block, although it still returns a level related error for some reason. It's not in the actual level list, but it's been slightly modernized.

Removed the extra ground collision under the player.