Obstacles indicate they are interactable
Added a second shader (with help from Dungeon Martian) that hilites obstacles when within the interact zone. This hilite plays nicely with the existing cursor hilite.

Had to make an adjustment to fences which removes the Blender integration stuff, but added a script-note that might help me remember to fix it later.