Health of limbs converted to floats instead of int
By making the max and current health on limbs floats instead of integers, the body parts should now be more prepared to accept half-healing values. Technically, half-damage values as well.

The half healing/damaging is not yet setup, but fractional hitpoints are accountable for in the code (eg if the player has .5 hp being treated the same as if they have .0 hp).

New functions were created to keep match statements tidy in these fractional conditions (with regards to damage; healing doesn't need to match fractionally, since fractional healing is ignored).