Progress numbers work great! Music fixe
Music would get freed on restart, but now is in its own autoload so that doesn't happen. (Thanks to dungeon martian)

Major thanks all along the way to Justin S for helping me get the progress code good. The numbers work great now although the bar is a bit goofy.