Updated progress formula; powerup proximity change
Progression formula now is simplified to empathy * abductions, more or less. (Unless no empathy is detected)

Made the powerup menu's collision (that hides obstacles) is now a bit higher.

Made the hurt armature be auto visible at the beginning.

Empathy defaults to 0 now.

Added a debug tool that allows for adding a typed number to the score.