Menu aliens should be skinnable
Setup a system for all of the alien skins to be held and loaded.

Menu aliens have some code to be assigned a skin via their terrain block.

Also set up a way for the menu aliens to have the shadows disabled without necessitating a new material (done by disabling shadows in their ready functions)