Refactored is_enemy and repaired Game Over state
Abductees now have is_enemy and is_military bools. Most is_enemy references are changed to is_military, except for those not pertaining to empathy checks (ie spawning new humans who are supposed be enenmies; in which case, they're marked as both enemies and military).

This is for prep for innocent humans to potentially also become enemies.

Finally, I attempted and I think succeeded in fixing a bug where an await would try running on a freed scene tree on restart of the game after a game over if the player was holding onto something (grab_ended).

Repaired by not pausing the game on game over, ensuring that the grab ends as soon as the player dies, and delaying showing the menu itself.