Godot skin works! Attack ray broken
Ended up not needing the prep for the Godot skin.

The attack ray, while "working" now, abducts humans on hover instead of click.

Clearly due to the attack ray being moved out of the action button function (after being added to it during last week's refactor), so will need to be fixed or possibly separated for different attacking types. Performance considerations might need to be made.